% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sei.R
\name{sei}
\alias{sei}
\title{Implemement storage efficient indexing}
\usage{
sei(x, R)
}
\arguments{
\item{x}{Records in X_1 that share an agreement pattern}

\item{R}{Maximum number of labels to keep}
}
\value{
Reduced vector of record labels
}
\description{
Implemement storage efficient indexing
}
